REFACTOR: Remove helper that is no longer necessary
Don't actually need whitespace at the string level. This is because once
the string is properly formatted, any string that was once composed
entirely of whitespace has been formatted down to an empty string. That
is `string-split` removes any whitespace, `filter` removes any empty
strings, so if `string-join` acts on an empty list it just returns an
empty string (which is exactly what is needed at that point).